Refactored

+++ sources/Editor/GuiComponents/Input.cpp
@@ -116,12 +116,14 @@ namespace Core
 
         InputTextCallback_UserData cb_user_data;
         cb_user_data.Str = &_buffer;
-        ImGui::InputTextEx("", _placeholder.c_str(), _buffer.data(), _buffer.capacity() + 1, _size,
-                           flags, InputTextCallback, &cb_user_data,
+        ImGui::InputTextEx("", _placeholder.c_str(), _buffer.data(),
+                           static_cast<int>(_buffer.capacity() + 1), _size, flags,
+                           InputTextCallback, &cb_user_data,
                            [this](const char* newText)
                            {
                                onInput.trigger(newText);
                            });
+
 
         if (isRO)
         {
